Containers/StatisticsInterface/requirements.txt
streamlit==1.39.0
streamlit-js-eval==0.1.7
extra-streamlit-components==0.1.71
python-dotenv